backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Antigravity"}
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy>=2.0.30",
    "pydantic-settings>=2.2.1",
    "openai>=1.30.1",
    "python-multipart>=0.0.26",
    "psycopg2-binary>=2.9.12",
    "passlib>=1.7.4",
    "bcrypt>=5.0.0",
    "python-jose[cryptography]>=3.5.0",
    "lxml>=6.1.0",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.pdm]
distribution = false

[dependency-groups]
dev = [
    "pytest>=8.2.0",
    "ruff>=0.4.4",
    "httpx>=0.28.1",
]
